UM-56 getDouble used instead of getInt

+++ src/main/java/com/archisacadeny/course/CourseRepository.java
@@ -185,7 +185,7 @@ public class CourseRepository {
             ResultSet rs = statement.getResultSet();
 
             while (rs.next()) {
-                grade = rs.getInt("sum");
+                grade = rs.getDouble("sum");
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
